changed versions to output itrations/sec

+++ gauss_seidel.h
@@ -5,6 +5,6 @@
 #define _GAUSS_SEIDEL_H
 
 // define your function prototype here
-int gauss_seidel(double ***u, double ***f, int N, int max_iter, double *threshold);
+double gauss_seidel(double ***u, double ***f, int N, int max_iter, double *threshold);
 
 #endif
